HALTimer fixes.
A number of things were fixed:
- one-shot did not work
- HAL_Time_Sleep_MicroSeconds broke the timers
- long handlers broke the timers
- Starting a timer broke the timers (for some reason CreateTimer started the timer)

Additional HALTimer fixes.

The active timers in the timer queue are now all adjusted properly before the next timer is inserted (and thus the new timer is not adjusted incorrectly).

+++ MicroFrameworkPK_v4_3/DeviceCode/Targets/Native/STM32F10x/DeviceCode/drivers/advancedtim/netmf_advancedtimer.h
@@ -78,7 +78,7 @@ public:
 	// The advanced timer combination will always be TIM1 -> TIM2.
 	DeviceStatus Initialize(UINT32 Prescaler, HAL_CALLBACK_FPN ISR, void* ISR_Param);
 
-	DeviceStatus SetCompare(UINT32 counterCorrection, UINT32 compareValue, SetCompareType scType);
+	DeviceStatus SetCompare(UINT64 counterCorrection, UINT32 compareValue, SetCompareType scType);
 
 	UINT32 GetCompare();
 	UINT32 GetCounter();
